# Reduce allocs by pooling gzip.Writer objects